+++ config-workspace/ThemeProvider/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rndna/theme-provider
 
+## 0.0.3
+
+### Patch Changes
+
+- update packages and stylings
+
 ## 0.0.2
 
 ### Patch Changes