+++ config-workspace/theme-config/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rndna/theme-config
 
+## 0.0.3
+
+### Patch Changes
+
+- update packages and stylings
+- Updated dependencies
+  - @rndna/base_style@0.0.6
+  - @rndna/fonts@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes