+++ packages-workspace/avatar_group/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rndna/avatar-group
 
+## 0.0.2
+
+### Patch Changes
+
+- update packages and stylings
+- Updated dependencies
+  - @rndna/icon@0.0.6
+
 ## 0.0.1
 
 ### Patch Changes