+++ packages-workspace/button/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @repo/button
 
+## 0.0.18
+
+### Patch Changes
+
+- update packages and stylings
+- Updated dependencies
+  - @rndna/icon@0.0.6
+  - @rndna/text@0.0.3
+
 ## 0.0.17
 
 ### Patch Changes