+++ packages-workspace/collapsible/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @repo/textfield
 
+## 0.0.2
+
+### Patch Changes
+
+- update packages and stylings
+- Updated dependencies
+  - @rndna/icon@0.0.6
+  - @rndna/text@0.0.3
+
 ## 0.0.1
 
 ### Patch Changes