+++ packages-workspace/drop-wrapper/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rndna/drop-wrapper
 
+## 0.0.2
+
+### Patch Changes
+
+- update packages and stylings
+- Updated dependencies
+  - @rndna/text@0.0.3
+
 ## 0.0.1
 
 ### Patch Changes