+++ packages-workspace/dropdown/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rndna/dropdown
 
+## 0.0.2
+
+### Patch Changes
+
+- update packages and stylings
+- Updated dependencies
+  - @rndna/fonts@0.0.3
+  - @rndna/text@0.0.3
+
 ## 0.0.1
 
 ### Patch Changes