+++ packages-workspace/fab/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rndna/fab
 
+## 0.0.4
+
+### Patch Changes
+
+- update packages and stylings
+- Updated dependencies
+  - @rndna/button@0.0.18
+  - @rndna/icon@0.0.6
+  - @rndna/text@0.0.3
+
 ## 0.0.3
 
 ### Patch Changes