+++ packages-workspace/hooks/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rndna/hooks
 
+## 0.0.3
+
+### Patch Changes
+
+- update packages and stylings
+- Updated dependencies
+  - @rndna/base_style@0.0.6
+
 ## 0.0.2
 
 ### Patch Changes