+++ packages-workspace/icon_button/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rndna/icon_button
 
+## 0.0.2
+
+### Patch Changes
+
+- update packages and stylings
+- Updated dependencies
+  - @rndna/button@0.0.18
+
 ## 0.0.1
 
 ### Patch Changes