+++ packages-workspace/image/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rndna/image
 
+## 0.0.3
+
+### Patch Changes
+
+- update packages and stylings
+- Updated dependencies
+  - @rndna/badge@0.0.3
+  - @rndna/icon@0.0.6
+  - @rndna/text@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes