+++ packages-workspace/menu/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rndna/menu
 
+## 0.0.2
+
+### Patch Changes
+
+- update packages and stylings
+- Updated dependencies
+  - @rndna/theme-provider@0.0.3
+  - @rndna/icon@0.0.6
+  - @rndna/text@0.0.3
+
 ## 0.0.1
 
 ### Patch Changes