+++ packages-workspace/text/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @repo/textfield
 
+## 0.0.3
+
+### Patch Changes
+
+- update packages and stylings
+- Updated dependencies
+  - @rndna/fonts@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes